requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
GitPython==3.1.40
networkx==3.2
requests==2.31.0
